Refactor service package structure: move services to appropriate sub-packages for better organization
