+++ src/main/java/com/ticketly/mseventseating/controller/OrganizationController.java
@@ -4,8 +4,8 @@ import com.ticketly.mseventseating.dto.organization.InviteStaffRequest;
 import com.ticketly.mseventseating.dto.organization.OrganizationMemberResponse;
 import com.ticketly.mseventseating.dto.organization.OrganizationRequest;
 import com.ticketly.mseventseating.dto.organization.OrganizationResponse;
-import com.ticketly.mseventseating.service.OrganizationService;
-import com.ticketly.mseventseating.service.OrganizationStaffService;
+import com.ticketly.mseventseating.service.organization.OrganizationService;
+import com.ticketly.mseventseating.service.organization.OrganizationStaffService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
