+++ src/main/java/com/ticketly/mseventseating/service/event/EventCreationService.java
@@ -6,9 +6,9 @@ import com.ticketly.mseventseating.exception.BadRequestException;
 import com.ticketly.mseventseating.factory.EventFactory;
 import com.ticketly.mseventseating.model.*;
 import com.ticketly.mseventseating.repository.EventRepository;
-import com.ticketly.mseventseating.service.LimitService;
-import com.ticketly.mseventseating.service.OrganizationService;
-import com.ticketly.mseventseating.service.S3StorageService;
+import com.ticketly.mseventseating.service.limts.LimitService;
+import com.ticketly.mseventseating.service.organization.OrganizationService;
+import com.ticketly.mseventseating.service.storage.S3StorageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import model.EventStatus;
