+++ src/main/java/com/ticketly/mseventseating/service/event/EventQueryService.java
@@ -5,8 +5,8 @@ import com.ticketly.mseventseating.dto.event.*;
 import com.ticketly.mseventseating.exception.ResourceNotFoundException;
 import com.ticketly.mseventseating.model.*;
 import com.ticketly.mseventseating.repository.EventRepository;
-import com.ticketly.mseventseating.service.OrganizationOwnershipService;
-import com.ticketly.mseventseating.service.S3StorageService;
+import com.ticketly.mseventseating.service.organization.OrganizationOwnershipService;
+import com.ticketly.mseventseating.service.storage.S3StorageService;
 import dto.SessionSeatingMapDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
