+++ src/main/java/com/ticketly/mseventseating/service/session/SessionOwnershipService.java
@@ -1,4 +1,4 @@
-package com.ticketly.mseventseating.service.event;
+package com.ticketly.mseventseating.service.session;
 
 import com.ticketly.mseventseating.exception.ResourceNotFoundException;
 import com.ticketly.mseventseating.model.EventSession;
