+++ src/test/java/com/ticketly/mseventseating/service/event/EventCreationServiceTest.java
@@ -8,9 +8,9 @@ import com.ticketly.mseventseating.exception.BadRequestException;
 import com.ticketly.mseventseating.factory.EventFactory;
 import com.ticketly.mseventseating.model.*;
 import com.ticketly.mseventseating.repository.EventRepository;
-import com.ticketly.mseventseating.service.LimitService;
-import com.ticketly.mseventseating.service.OrganizationService;
-import com.ticketly.mseventseating.service.S3StorageService;
+import com.ticketly.mseventseating.service.limts.LimitService;
+import com.ticketly.mseventseating.service.organization.OrganizationService;
+import com.ticketly.mseventseating.service.storage.S3StorageService;
 import model.EventStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
