+++ src/test/java/com/ticketly/mseventseating/service/event/EventQueryServiceTest.java
@@ -5,8 +5,8 @@ import com.ticketly.mseventseating.dto.event.EventSummaryDTO;
 import com.ticketly.mseventseating.exception.ResourceNotFoundException;
 import com.ticketly.mseventseating.model.*;
 import com.ticketly.mseventseating.repository.EventRepository;
-import com.ticketly.mseventseating.service.OrganizationOwnershipService;
-import com.ticketly.mseventseating.service.S3StorageService;
+import com.ticketly.mseventseating.service.organization.OrganizationOwnershipService;
+import com.ticketly.mseventseating.service.storage.S3StorageService;
 import model.EventStatus;
 import model.SessionStatus;
 import org.junit.jupiter.api.BeforeEach;
